# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Hair Salon Booking API",
    "description": "API for Hair Salon Booking Application",
    "version": "1.0.0",
    "contact": {
      "name": "Developer",
      "email": "developer@hairsalon.com"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT Authorization header using Bearer scheme. Example: 'Bearer {token}'"
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/booking/create": {
      "post": {
        "tags": [
          "Booking"
        ],
        "summary": "Create new booking",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customerID": {
                    "type": "string"
                  },
                  "stylistID": {
                    "type": "string"
                  },
                  "serviceIDs": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "bookingDate": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Booking created"
          }
        }
      }
    },
...
    "/news/getDetail": {
      "get": {
        "tags": [
          "News"
        ],
        "summary": "Get news detail",
        "parameters": [
          {
            "in": "query",
            "name": "newsID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "News details"
          }
        }
      }
    },
    "/news/create": {
      "post": {
        "tags": [
          "News"
        ],
        "summary": "Create news",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "News created"
          }
        }
      }
    },
    "/news/update": {
      "patch": {
        "tags": [
          "News"
        ],
        "summary": "Update news",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "newsID": {
                    "type": "string"
                  },
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "News updated"
          }
        }
      }
    },
    "/news/delete": {
      "patch": {
        "tags": [
          "News"
        ],
        "summary": "Soft delete news",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "newsID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "News deleted"
          }
        }
      },
      "delete": {
        "tags": [
          "News"
        ],
        "summary": "Permanently delete news",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "newsID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "News deleted permanently"
          }
        }
      }
    },
...
  },
...
}